README.md update kotlinx.CLI changed to Clikt add while loop too prog
