+++ README.md
@@ -1,29 +1,33 @@
 # ConsoleWeatherApp
 ##  Описание
-Данный сервис позволяет получить текущее состояние погоды в выбранном городе
+Данный сервис позволяет получить текущее состояние погоды в выбранном городе.
 ## Использование
-На вход сервис ожидает команду следующего вида:
+Сервис может быть запущен со следующими параметрами:
 ```
-[city1, city2, ...] [options]
-
-Аргументы:
-    cities -> Название города, в котором требуется получить состояние погоды { String } (города с пробельным 
-    символом пишутся в кавычках)
-
 Options:
-    1. --unit, -u [default:METRIC] -> Выбор единицы измерения температуры { Значение должно быть одним из 
-    следующих [imperial(Фаренгейт), metric(Цельсий), standard(Кельвин)] }
-    2. --wind, -w [default:false] -> Вывод информации о ветре
-    3. --temperature, -t [default:false] -> Вывод дополнительной информации о температуре
-    4. --apiKey, -a [default:""] -> Изменение API ключа на другой { String }
-    5. --help, -h -> Инофрмация об использовании
+  -w, --wind                       Вывод информации о ветре
+  -t, --temp                       Вывод дополнительной информациии о температуре
+  -a, --api TEXT                   Сменить API ключ на заданный
+  -u, --unit [IMPERIAL|METRIC|STANDARD]
+                                   Сменить единицу измерения температуры
+  -h, --help                       Вывод сообщения со всеми возможными параметрыми
 ```
-Пример вызова команды:
+Единицы измерения температуры можно интерпретировать следующим образом:
+* IMPERIAL - Фаренгейт
+* METRIC - Цельсий
+* STANDARD - Кельвин
+
+После того как пользователь запустил сервис с различными параметрами, он может начать с ним работать.
+Для этого ему следует ввести название любого города. Если городов несколько, то их требуется вводить
+через ",".
+
+Пример запуска сервиса:
 ```
-МосквА НеГород Нижневартовск -u standard -w -t
+-u standard -w
 ```
-Результат:
+Пример пользовательского ввода и результата:
 ```
+user input: Москва
 Москва
 ---------Общая информация---------
 Широта: 37.6156
@@ -35,6 +39,7 @@ Options:
 Направление ветра: Западный
 ----------------------------------
 
+user input: НеГород, Нижневартовск
 Такой город не найден: НеГород
 
 Нижневартовск
@@ -55,5 +60,5 @@ docker build -t [image_name] https://github.com/dVede/ConsoleWeatherApp.git
 ```
 Запустить контейнер:
 ```
-docker run [image_name] [args]
+docker run [image_name] [OPTIONS]
 ```
