+++ src/main/kotlin/UnitsTemp.kt
@@ -1,0 +1,7 @@
+package com.dvede.weather
+
+enum class UnitsTemp(val str: String) {
+    IMPERIAL("°F"),
+    METRIC( "°C"),
+    STANDARD("K")
+}
